Create interface for repository
