+++ app/src/main/java/com/yorkismine/expenseapp/model/ExpenseRepository.java
@@ -1,121 +1,24 @@
 package com.yorkismine.expenseapp.model;
 
-import android.app.Application;
-import android.content.Context;
-import android.os.AsyncTask;
-import android.widget.EditText;
-
 import androidx.lifecycle.LiveData;
 
 import java.util.List;
 
-public class ExpenseRepository {
-    private LiveData<List<Expense>> allExpenses;
-    private LiveData<List<Expense>> sortByDate;
-    private LiveData<List<Expense>> sortByName;
-    private LiveData<List<Expense>> sortBySum;
-    private ExpenseDAO expenseDAO;
-    private ExpenseRepository instance;
+public interface ExpenseRepository {
 
-    public ExpenseRepository(Context context){
-        ExpenseDatabase database = ExpenseDatabase.getInstance(context);
-        expenseDAO = database.expenseDAO();
-        allExpenses = expenseDAO.getAllExpenses();
-        sortByDate = expenseDAO.sortByDate();
-        sortByName = expenseDAO.sortByName();
-        sortBySum = expenseDAO.sortBySum();
-    }
+    void insert(Expense expense);
 
-    public void insert(Expense expense){
-        new InsertExpenseAsyncTask(expenseDAO).execute(expense);
-    }
+    void update(Expense expense);
 
-    public void update(Expense expense){
-        new UpdateExpenseAsyncTask(expenseDAO).execute(expense);
-    }
+    void delete(Expense expense);
 
-    public void delete(Expense expense){
-        new DeleteExpenseAsyncTask(expenseDAO).execute(expense);
-    }
+    void deleteAllExpenses();
 
-    public void deleteAllExpenses(){
-        new DeleteAllExpensesAsyncTask(expenseDAO).execute();
-    }
+    LiveData<List<Expense>> getAllExpenses();
 
-    public LiveData<List<Expense>> getAllExpenses(){
-        return allExpenses;
-    }
+    LiveData<List<Expense>> sortByDate();
 
-    public LiveData<List<Expense>> sortByDate(){
-        return sortByDate;
-    }
-    public LiveData<List<Expense>> sortByName(){
-        return sortByName;
-    }
-    public LiveData<List<Expense>> sortBySum(){
-        return sortBySum;
-    }
+    LiveData<List<Expense>> sortByName();
 
-
-
-
-    private static class InsertExpenseAsyncTask extends AsyncTask<Expense, Void, Void>{
-        private ExpenseDAO expenseDAO;
-
-        public InsertExpenseAsyncTask(ExpenseDAO expenseDAO) {
-            this.expenseDAO = expenseDAO;
-        }
-
-        @Override
-        protected Void doInBackground(Expense... expenses) {
-            expenseDAO.insert(expenses[0]);
-            return null;
-        }
-    }
-
-    private static class UpdateExpenseAsyncTask extends AsyncTask<Expense, Void, Void>{
-        private ExpenseDAO expenseDAO;
-
-        public UpdateExpenseAsyncTask(ExpenseDAO expenseDAO) {
-            this.expenseDAO = expenseDAO;
-        }
-
-        @Override
-        protected Void doInBackground(Expense... expenses) {
-            expenseDAO.update(expenses[0]);
-            return null;
-        }
-    }
-
-    private static class DeleteExpenseAsyncTask extends AsyncTask<Expense, Void, Void>{
-        private ExpenseDAO expenseDAO;
-
-        public DeleteExpenseAsyncTask(ExpenseDAO expenseDAO) {
-            this.expenseDAO = expenseDAO;
-        }
-
-        @Override
-        protected Void doInBackground(Expense... expenses) {
-            expenseDAO.delete(expenses[0]);
-            return null;
-        }
-    }
-
-    private static class DeleteAllExpensesAsyncTask extends AsyncTask<Void, Void, Void>{
-        private ExpenseDAO expenseDAO;
-
-        public DeleteAllExpensesAsyncTask(ExpenseDAO expenseDAO) {
-            this.expenseDAO = expenseDAO;
-        }
-
-        @Override
-        protected Void doInBackground(Void... voids) {
-            expenseDAO.deleteAllExpenses();
-            return null;
-        }
-    }
-
-
-
-
+    LiveData<List<Expense>> sortBySum();
 }
