+++ app/src/main/java/com/yorkismine/expenseapp/ui/statistics/StatisticsViewModel.java
@@ -6,6 +6,7 @@ import androidx.lifecycle.AndroidViewModel;
 import androidx.lifecycle.LiveData;
 
 import com.yorkismine.expenseapp.model.Expense;
+import com.yorkismine.expenseapp.model.ExpenseRepositoryImpl;
 import com.yorkismine.expenseapp.model.ExpenseRepository;
 
 import java.util.List;
@@ -17,7 +18,7 @@ public class StatisticsViewModel extends AndroidViewModel {
 
     public StatisticsViewModel(Application application) {
         super(application);
-        repository = new ExpenseRepository(application);
+        repository = new ExpenseRepositoryImpl(application);
         allExpenses = repository.getAllExpenses();
     }
